backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
motor>=3.3.0
cloudinary>=1.36.0
python-multipart>=0.0.6
